requirements.txt
streamlit>=1.40.0
httpx>=0.28.0
google-genai>=1.10.0
fal-client>=0.5.0
cloudinary>=1.41.0
gspread>=6.1.0
google-auth>=2.35.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0
pydantic>=2.10.0
